Uncommenting feedback commands
Now that we have solved the issue behind receiving feedback from the Pololu Motor Drivers (we modified the command POLOLUCONTROL_READ_FEEDBACK_COMMAND to be A7 instead of A5), we would like to receive feedback from the elbow and forearm now.  We will work with the wrist once the wrist is working.